Use string interpolation instead of concatenation, don't forward failed parses.
Using string interpolation is safer here because it calls to_s on each of the
values being inserted into the string. It's also allegedly faster than string
concatenation although I don't have anything to back that up here.

We've seen errors from this plugin when the JSON codec fails to parse a
message, since it sends some unparsed fragments as an array via
event['message']. This patch attempts to skip forwarding such fragments by
looking for a '_jsonparsefailure' tag but the switch to string interpolation
will keep an exception from being raised even if an unexpected data type is
passed.